Allow actions to be persistent or not.
On previous implementation all the actions are executed one time only,
since it used a checkpoint file to avoid double execution.

That practice is good for actions that affect files in the structure,
but render useless for tasks that only alter current environment, like
compiler flags or environment.

This new method allows execution of persistent tasks (:persist => true)
and also non-persisting tasks.

  after :extract do
    # this will be executed always
  end

  after :extract, :persist => true do
    # this only will be executed once and marked in the checkpoint
  end