Added Zipfile extraction to Utils.
Leverage in RubyZip (0.9.1) to extract zip file contents into
specified destination folders.

RubyZip source was added to lib/contrib/zip folder.

Defined Zip::extract as simple helper (hides the complexity of RubyZip).

Updated OneClick::Utils#extract to reflect the dependency

Generate an error if any other file type is tried to be
extracted (we need to bootstrap the the extraction).